backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
python-telegram-bot>=20.7
djangorestframework-simplejwt>=5.3
django-celery-beat>=2.5
Pillow>=10.0
requests>=2.31
dj-database-url>=2.1
whitenoise>=6.6
gunicorn>=21.2